+++ tortoise/contrib/test/__init__.py
@@ -3,7 +3,8 @@ import os as _os
 import unittest
 from asyncio.events import AbstractEventLoop
 from functools import wraps
-from typing import Any, Callable, List, Optional
+from types import ModuleType
+from typing import Any, Callable, Iterable, List, Optional, Union
 from unittest import SkipTest, expectedFailure, skip, skipIf, skipUnless
 from unittest.result import TestResult
 
@@ -11,7 +12,7 @@ from asynctest import TestCase as _TestCase
 from asynctest import _fail_on
 from asynctest.case import _Policy
 
-from tortoise import Tortoise
+from tortoise import Model, Tortoise
 from tortoise.backends.base.config_generator import generate_config as _generate_config
 from tortoise.exceptions import DBConnectionError
 from tortoise.transactions import current_transaction_map
@@ -45,11 +46,11 @@ _CONFIG: dict = {}
 _CONNECTIONS: dict = {}
 _SELECTOR = None
 _LOOP: AbstractEventLoop = None  # type: ignore
-_MODULES: List[str] = []
+_MODULES: Iterable[Union[str, ModuleType]] = []
 _CONN_MAP: dict = {}
 
 
-def getDBConfig(app_label: str, modules: List[str]) -> dict:
+def getDBConfig(app_label: str, modules: Iterable[Union[str, ModuleType]]) -> dict:
     """
     DB Config factory, for use in testing.
 
@@ -84,7 +85,7 @@ def _restore_default() -> None:
 
 
 def initializer(
-    modules: List[str],
+    modules: Iterable[Union[str, ModuleType]],
     db_url: Optional[str] = None,
     app_label: str = "models",
     loop: Optional[AbstractEventLoop] = None,
@@ -283,6 +284,14 @@ class SimpleTestCase(_TestCase):  # type: ignore
             # clear the outcome, no more needed
             self._outcome = None
 
+    def assertListSortEqual(self, list1: List[Any], list2: List[Any], msg: Any = ...) -> None:
+        if isinstance(list1[0], Model):
+            super().assertListEqual(
+                sorted(list1, key=lambda x: x.pk), sorted(list2, key=lambda x: x.pk)
+            )
+        elif isinstance(list1[0], tuple):
+            super().assertListEqual(sorted(list1), sorted(list2))
+
     async def _run_outcome(self, outcome, expecting_failure: bool, testMethod: Callable) -> None:
         with outcome.testPartExecutor(self):
             await self._setUp()
@@ -318,7 +327,7 @@ class IsolatedTestCase(SimpleTestCase):
     If you define a ``tortoise_test_modules`` list, it overrides the DB setup module for the tests.
     """
 
-    tortoise_test_modules: List[str] = []
+    tortoise_test_modules: Iterable[Union[str, ModuleType]] = []
 
     async def _setUpDB(self) -> None:
         config = getDBConfig(app_label="models", modules=self.tortoise_test_modules or _MODULES)