+++ tortoise/models.py
@@ -46,7 +46,10 @@ from tortoise.fields.relational import (
 )
 from tortoise.filters import get_filters_for_field
 from tortoise.functions import Function
+from tortoise.indexes import Index
+from tortoise.manager import Manager
 from tortoise.queryset import ExistsQuery, Q, QuerySet, QuerySetSingle
+from tortoise.router import router
 from tortoise.signals import Signals
 from tortoise.transactions import current_transaction_map, in_transaction
 
@@ -182,6 +185,7 @@ class MetaInfo:
         "basetable",
         "_filters",
         "unique_together",
+        "manager",
         "indexes",
         "pk_attr",
         "generated_db_fields",
@@ -198,6 +202,7 @@ class MetaInfo:
 
     def __init__(self, meta: "Model.Meta") -> None:
         self.abstract: bool = getattr(meta, "abstract", False)
+        self.manager: Manager = getattr(meta, "manager", Manager())
         self.db_table: str = getattr(meta, "table", "")
         self.app: Optional[str] = getattr(meta, "app", None)
         self.unique_together: Tuple[Tuple[str, ...], ...] = get_together(meta, "unique_together")
@@ -609,8 +614,11 @@ class ModelMeta(type):
 
         if new_class.__doc__ and not meta.table_description:
             meta.table_description = inspect.cleandoc(new_class.__doc__).split("\n")[0]
-
+        for key, value in attrs.items():
+            if isinstance(value, Manager):
+                value._model = new_class
         meta._model = new_class
+        meta.manager._model = new_class
         meta.finalise_fields()
         return new_class
 
@@ -727,6 +735,10 @@ class Model(metaclass=ModelMeta):
         if not self.pk:
             raise TypeError("Model instances without id are unhashable")
         return hash(self.pk)
+
+    def __iter__(self):
+        for field in self._meta.fields_map:
+            yield field, getattr(self, field)
 
     def __eq__(self, other: object) -> bool:
         return type(other) is type(self) and self.pk == other.pk  # type: ignore
@@ -880,10 +892,10 @@ class Model(metaclass=ModelMeta):
         :param force_create: Forces creation of the record
         :param force_update: Forces updating of the record
 
-        :raises IncompleteInstanceError: If the model is partial and the fields are not available for persistance.
+        :raises IncompleteInstanceError: If the model is partial and the fields are not available for persistence.
         :raises IntegrityError: If the model can't be created or updated (specifically if force_create or force_update has been set)
         """
-        db = using_db or self._meta.db
+        db = using_db or self._choose_db(True)
         executor = db.executor_class(model=self.__class__, db=db)
         if self._partial:
             if update_fields:
@@ -900,7 +912,7 @@ class Model(metaclass=ModelMeta):
                 raise IncompleteInstanceError(
                     f"{self.__class__.__name__} is a partial model, can only be saved with the relevant update_field provided"
                 )
-        await self._pre_save(using_db, update_fields)
+        await self._pre_save(db, update_fields)
 
         if force_create:
             await executor.execute_insert(self)
@@ -924,7 +936,7 @@ class Model(metaclass=ModelMeta):
                 created = True
 
         self._saved_in_db = True
-        await self._post_save(using_db, created, update_fields)
+        await self._post_save(db, created, update_fields)
 
     async def delete(self, using_db: Optional[BaseDBAsyncClient] = None) -> None:
         """
@@ -934,12 +946,12 @@ class Model(metaclass=ModelMeta):
 
         :raises OperationalError: If object has never been persisted.
         """
-        db = using_db or self._meta.db
+        db = using_db or self._choose_db(True)
         if not self._saved_in_db:
             raise OperationalError("Can't delete unpersisted record")
-        await self._pre_delete(using_db)
+        await self._pre_delete(db)
         await db.executor_class(model=self.__class__, db=db).execute_delete(self)
-        await self._post_delete(using_db)
+        await self._post_delete(db)
 
     async def fetch_related(self, *args: Any, using_db: Optional[BaseDBAsyncClient] = None) -> None:
         """
@@ -952,7 +964,7 @@ class Model(metaclass=ModelMeta):
         :param args: The related fields that should be fetched.
         :param using_db: Specific DB connection to use instead of default bound
         """
-        db = using_db or self._meta.db
+        db = using_db or self._choose_db()
         await db.executor_class(model=self.__class__, db=db).fetch_for_list([self], *args)
 
     async def refresh_from_db(
@@ -972,11 +984,25 @@ class Model(metaclass=ModelMeta):
         """
         if not self._saved_in_db:
             raise OperationalError("Can't refresh unpersisted record")
-        qs = QuerySet(self.__class__).only(*(fields or []))
-        using_db and qs.using_db(using_db)
+        db = using_db or self._choose_db()
+        qs = QuerySet(self.__class__).using_db(db).only(*(fields or []))
         obj = await qs.get(pk=self.pk)
         for field in fields or self._meta.fields_map:
             setattr(self, field, getattr(obj, field, None))
+
+    @classmethod
+    def _choose_db(cls, for_write: bool = False):
+        """
+        Return the connection that will be used if this query is executed now.
+
+        :param for_write: Whether this query for write.
+        :return: BaseDBAsyncClient:
+        """
+        if for_write:
+            db = router.db_for_write(cls)
+        else:
+            db = router.db_for_read(cls)
+        return db or cls._meta.db
 
     @classmethod
     async def get_or_create(
@@ -995,7 +1021,7 @@ class Model(metaclass=ModelMeta):
         """
         if not defaults:
             defaults = {}
-        db = using_db if using_db else cls._meta.db
+        db = using_db or cls._choose_db(True)
         async with in_transaction(connection_name=db.connection_name):
             instance = await cls.filter(**kwargs).first()
             if instance:
@@ -1024,7 +1050,7 @@ class Model(metaclass=ModelMeta):
         """
         if not defaults:
             defaults = {}
-        db = using_db if using_db else cls._meta.db
+        db = using_db or cls._choose_db(True)
         async with in_transaction(connection_name=db.connection_name):
             instance = await cls.filter(**kwargs).first()
             if instance:
@@ -1058,7 +1084,7 @@ class Model(metaclass=ModelMeta):
         """
         instance = cls(**kwargs)
         instance._saved_in_db = False
-        db = kwargs.get("using_db") or cls._meta.db
+        db = kwargs.get("using_db") or cls._choose_db(True)
         await instance.save(using_db=db, force_create=True)
         return instance
 
@@ -1093,7 +1119,7 @@ class Model(metaclass=ModelMeta):
         :param batch_size: How many objects are created in a single query
         :param using_db: Specific DB connection to use instead of default bound
         """
-        db = using_db or cls._meta.db
+        db = using_db or cls._choose_db(True)
         await db.executor_class(model=cls, db=db).execute_bulk_insert(objects, batch_size)
 
     @classmethod
@@ -1101,7 +1127,7 @@ class Model(metaclass=ModelMeta):
         """
         Generates a QuerySet that returns the first record.
         """
-        return QuerySet(cls).first()
+        return cls._meta.manager.get_queryset().first()
 
     @classmethod
     def filter(cls: Type[MODEL], *args: Q, **kwargs: Any) -> QuerySet[MODEL]:
@@ -1111,7 +1137,7 @@ class Model(metaclass=ModelMeta):
         :param args: Q funtions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        return QuerySet(cls).filter(*args, **kwargs)
+        return cls._meta.manager.get_queryset().filter(*args, **kwargs)
 
     @classmethod
     def exclude(cls: Type[MODEL], *args: Q, **kwargs: Any) -> QuerySet[MODEL]:
@@ -1121,7 +1147,7 @@ class Model(metaclass=ModelMeta):
         :param args: Q funtions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        return QuerySet(cls).exclude(*args, **kwargs)
+        return cls._meta.manager.get_queryset().exclude(*args, **kwargs)
 
     @classmethod
     def annotate(cls: Type[MODEL], **kwargs: Union[Function, Term]) -> QuerySet[MODEL]:
@@ -1130,14 +1156,14 @@ class Model(metaclass=ModelMeta):
 
         :param kwargs: Parameter name and the Function/Aggregation to annotate with.
         """
-        return QuerySet(cls).annotate(**kwargs)
+        return cls._meta.manager.get_queryset().annotate(**kwargs)
 
     @classmethod
     def all(cls: Type[MODEL]) -> QuerySet[MODEL]:
         """
         Returns the complete QuerySet.
         """
-        return QuerySet(cls)
+        return cls._meta.manager.get_queryset()
 
     @classmethod
     def get(cls: Type[MODEL], *args: Q, **kwargs: Any) -> QuerySetSingle[MODEL]:
@@ -1154,7 +1180,7 @@ class Model(metaclass=ModelMeta):
         :raises MultipleObjectsReturned: If provided search returned more than one object.
         :raises DoesNotExist: If object can not be found.
         """
-        return QuerySet(cls).get(*args, **kwargs)
+        return cls._meta.manager.get_queryset().get(*args, **kwargs)
 
     @classmethod
     def exists(cls: Type[MODEL], *args: Q, **kwargs: Any) -> ExistsQuery:
@@ -1168,7 +1194,7 @@ class Model(metaclass=ModelMeta):
         :param args: Q funtions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        return QuerySet(cls).filter(*args, **kwargs).exists()
+        return cls._meta.manager.get_queryset().filter(*args, **kwargs).exists()
 
     @classmethod
     def get_or_none(cls: Type[MODEL], *args: Q, **kwargs: Any) -> QuerySetSingle[Optional[MODEL]]:
@@ -1177,12 +1203,12 @@ class Model(metaclass=ModelMeta):
 
         .. code-block:: python3
 
-            user = await User.get(username="foo")
+            user = await User.get_or_none(username="foo")
 
         :param args: Q funtions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        return QuerySet(cls).get_or_none(*args, **kwargs)
+        return cls._meta.manager.get_queryset().get_or_none(*args, **kwargs)
 
     @classmethod
     async def fetch_for_list(
@@ -1198,7 +1224,7 @@ class Model(metaclass=ModelMeta):
         :param args: Relation names to fetch.
         :param using_db: DO NOT USE
         """
-        db = using_db or cls._meta.db
+        db = using_db or cls._choose_db()
         await db.executor_class(model=cls, db=db).fetch_for_list(instance_list, *args)
 
     @classmethod
@@ -1222,12 +1248,14 @@ class Model(metaclass=ModelMeta):
         if not isinstance(_together, (tuple, list)):
             raise ConfigurationError(f"'{cls.__name__}.{together}' must be a list or tuple.")
 
-        if any(not isinstance(unique_fields, (tuple, list)) for unique_fields in _together):
+        if any(not isinstance(unique_fields, (tuple, list, Index)) for unique_fields in _together):
             raise ConfigurationError(
                 f"All '{cls.__name__}.{together}' elements must be lists or tuples."
             )
 
         for fields_tuple in _together:
+            if isinstance(fields_tuple, Index):
+                fields_tuple = fields_tuple.fields
             for field_name in fields_tuple:
                 field = cls._meta.fields_map.get(field_name)