+++ tortoise/queryset.py
@@ -41,8 +41,10 @@ from tortoise.fields.relational import (
 )
 from tortoise.functions import Function
 from tortoise.query_utils import Prefetch, Q, QueryModifier, _get_joins_for_related_field
+from tortoise.router import router
 
 # Empty placeholder - Should never be edited.
+
 QUERY: QueryBuilder = QueryBuilder()
 
 if TYPE_CHECKING:  # pragma: nocoverage
@@ -87,6 +89,20 @@ class AwaitableQuery(Generic[MODEL]):
         self._db: BaseDBAsyncClient = None  # type: ignore
         self.capabilities: Capabilities = model._meta.db.capabilities
         self._annotations: Dict[str, Function] = {}
+
+    def _choose_db(self, for_write: bool = False) -> BaseDBAsyncClient:
+        """
+        Return the connection that will be used if this query is executed now.
+
+        :return: BaseDBAsyncClient:
+        """
+        if self._db:
+            return self._db
+        if for_write:
+            db = router.db_for_write(self.model)
+        else:
+            db = router.db_for_read(self.model)
+        return db or self.model._meta.db
 
     def resolve_filters(
         self,
@@ -255,8 +271,13 @@ class QuerySet(AwaitableQuery[MODEL]):
         "_custom_filters",
         "_group_bys",
         "_select_for_update",
+        "_select_for_update_nowait",
+        "_select_for_update_skip_locked",
+        "_select_for_update_of",
         "_select_related",
         "_select_related_idx",
+        "_use_indexes",
+        "_force_indexes",
     )
 
     def __init__(self, model: Type[MODEL]) -> None:
@@ -277,10 +298,15 @@ class QuerySet(AwaitableQuery[MODEL]):
         self._fields_for_select: Tuple[str, ...] = ()
         self._group_bys: Tuple[str, ...] = ()
         self._select_for_update: bool = False
+        self._select_for_update_nowait: bool = False
+        self._select_for_update_skip_locked: bool = False
+        self._select_for_update_of: Set[str] = set()
         self._select_related: Set[str] = set()
         self._select_related_idx: List[
             Tuple["Type[Model]", int, str, "Type[Model]"]
         ] = []  # format with: model,idx,model_name,parent_model
+        self._force_indexes: Set[str] = set()
+        self._use_indexes: Set[str] = set()
 
     def _clone(self) -> "QuerySet[MODEL]":
         queryset = QuerySet.__new__(QuerySet)
@@ -306,8 +332,13 @@ class QuerySet(AwaitableQuery[MODEL]):
         queryset._custom_filters = copy(self._custom_filters)
         queryset._group_bys = copy(self._group_bys)
         queryset._select_for_update = self._select_for_update
+        queryset._select_for_update_nowait = self._select_for_update_nowait
+        queryset._select_for_update_skip_locked = self._select_for_update_skip_locked
+        queryset._select_for_update_of = self._select_for_update_of
         queryset._select_related = self._select_related
         queryset._select_related_idx = self._select_related_idx
+        queryset._force_indexes = self._force_indexes
+        queryset._use_indexes = self._use_indexes
         return queryset
 
     def _filter_or_exclude(self, *args: Q, negate: bool, **kwargs: Any) -> "QuerySet[MODEL]":
@@ -411,7 +442,9 @@ class QuerySet(AwaitableQuery[MODEL]):
         queryset._distinct = True
         return queryset
 
-    def select_for_update(self) -> "QuerySet[MODEL]":
+    def select_for_update(
+        self, nowait: bool = False, skip_locked: bool = False, of: Tuple[str, ...] = ()
+    ) -> "QuerySet[MODEL]":
         """
         Make QuerySet select for update.
 
@@ -421,10 +454,13 @@ class QuerySet(AwaitableQuery[MODEL]):
         if self.capabilities.support_for_update:
             queryset = self._clone()
             queryset._select_for_update = True
+            queryset._select_for_update_nowait = nowait
+            queryset._select_for_update_skip_locked = skip_locked
+            queryset._select_for_update_of = set(of)
             return queryset
         return self
 
-    def annotate(self, **kwargs: Function) -> "QuerySet[MODEL]":
+    def annotate(self, **kwargs: Union[Function, Term]) -> "QuerySet[MODEL]":
         """
         Annotate result with aggregation or function result.
 
@@ -534,6 +570,8 @@ class QuerySet(AwaitableQuery[MODEL]):
             q_objects=self._q_objects,
             annotations=self._annotations,
             custom_filters=self._custom_filters,
+            limit=self._limit,
+            orderings=self._orderings,
         )
 
     def update(self, **kwargs: Any) -> "UpdateQuery":
@@ -653,6 +691,29 @@ class QuerySet(AwaitableQuery[MODEL]):
             queryset._select_related.add(field)
         return queryset
 
+    def force_index(self, *index_names: str) -> "QuerySet[MODEL]":
+        """
+        The FORCE INDEX hint acts like USE INDEX (index_list),
+        with the addition that a table scan is assumed to be very expensive.
+        """
+        if self.capabilities.support_index_hint:
+            queryset = self._clone()
+            for index_name in index_names:
+                queryset._force_indexes.add(index_name)
+            return queryset
+        return self
+
+    def use_index(self, *index_names: str) -> "QuerySet[MODEL]":
+        """
+        The USE INDEX (index_list) hint tells MySQL to use only one of the named indexes to find rows in the table.
+        """
+        if self.capabilities.support_index_hint:
+            queryset = self._clone()
+            for index_name in index_names:
+                queryset._use_indexes.add(index_name)
+            return queryset
+        return self
+
     def prefetch_related(self, *args: Union[str, Prefetch]) -> "QuerySet[MODEL]":
         """
         Like ``.fetch_related()`` on instance, but works on all objects in QuerySet.
@@ -699,7 +760,7 @@ class QuerySet(AwaitableQuery[MODEL]):
             **The output format may (and will) vary greatly depending on the database backend.**
         """
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
         return await self._db.executor_class(model=self.model, db=self._db).execute_explain(
             self.query
@@ -726,9 +787,9 @@ class QuerySet(AwaitableQuery[MODEL]):
 
         table = self._join_table_by_field(table, field, field_object)
         related_fields = field_object.related_model._meta.db_fields
-        self._select_related_idx.append(
-            (field_object.related_model, len(related_fields), field, model)
-        )
+        append_item = (field_object.related_model, len(related_fields), field, model)
+        if append_item not in self._select_related_idx:
+            self._select_related_idx.append(append_item)
         for related_field in related_fields:
             self.query = self.query.select(
                 table[related_field].as_(f"{table.get_table_name()}.{related_field}")
@@ -748,9 +809,9 @@ class QuerySet(AwaitableQuery[MODEL]):
         self._select_related_idx = []
         table = self.model._meta.basetable
         if self._fields_for_select:
-            self._select_related_idx.append(
-                (self.model, len(self._fields_for_select), table, self.model)
-            )
+            append_item = (self.model, len(self._fields_for_select), table, self.model)
+            if append_item not in self._select_related_idx:
+                self._select_related_idx.append(append_item)
             db_fields_for_select = [
                 table[self.model._meta.fields_db_projection[field]].as_(field)
                 for field in self._fields_for_select
@@ -758,9 +819,9 @@ class QuerySet(AwaitableQuery[MODEL]):
             self.query = copy(self.model._meta.basequery).select(*db_fields_for_select)
         else:
             self.query = copy(self.model._meta.basequery_all_fields)
-            self._select_related_idx.append(
-                (self.model, len(self.model._meta.db_fields), table, self.model)
-            )
+            append_item = (self.model, len(self.model._meta.db_fields), table, self.model)
+            if append_item not in self._select_related_idx:
+                self._select_related_idx.append(append_item)
         self.resolve_ordering(
             self.model, self.model._meta.basetable, self._orderings, self._annotations
         )
@@ -777,7 +838,11 @@ class QuerySet(AwaitableQuery[MODEL]):
         if self._distinct:
             self.query._distinct = True
         if self._select_for_update:
-            self.query._for_update = True
+            self.query = self.query.for_update(
+                self._select_for_update_nowait,
+                self._select_for_update_skip_locked,
+                self._select_for_update_of,
+            )
         if self._select_related:
             for field in self._select_related:
                 field_split = field.split("__")
@@ -787,10 +852,14 @@ class QuerySet(AwaitableQuery[MODEL]):
                     field=field_split[0],
                     forwarded_fields="__".join(field_split[1:]) if len(field_split) > 1 else "",
                 )
+        if self._force_indexes:
+            self.query = self.query.force_index(*self._force_indexes)
+        if self._use_indexes:
+            self.query = self.query.use_index(*self._use_indexes)
 
     def __await__(self) -> Generator[Any, None, List[MODEL]]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db(self._select_for_update)  # type: ignore
         self._make_query()
         return self._execute().__await__()
 
@@ -876,7 +945,7 @@ class UpdateQuery(AwaitableQuery):
 
     def __await__(self) -> Generator[Any, None, int]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db(True)  # type: ignore
         self._make_query()
         return self._execute().__await__()
 
@@ -885,7 +954,7 @@ class UpdateQuery(AwaitableQuery):
 
 
 class DeleteQuery(AwaitableQuery):
-    __slots__ = ("q_objects", "annotations", "custom_filters")
+    __slots__ = ("q_objects", "annotations", "custom_filters", "orderings", "limit")
 
     def __init__(
         self,
@@ -894,15 +963,25 @@ class DeleteQuery(AwaitableQuery):
         q_objects: List[Q],
         annotations: Dict[str, Any],
         custom_filters: Dict[str, Dict[str, Any]],
+        limit: Optional[int],
+        orderings: List[Tuple[str, str]],
     ) -> None:
         super().__init__(model)
         self.q_objects = q_objects
         self.annotations = annotations
         self.custom_filters = custom_filters
         self._db = db
+        self.limit = limit
+        self.orderings = orderings
 
     def _make_query(self) -> None:
         self.query = copy(self.model._meta.basequery)
+        if self.capabilities.support_update_limit_order_by:
+            self.resolve_ordering(
+                self.model, self.model._meta.basetable, self.orderings, self.annotations
+            )
+            if self.limit:
+                self.query._limit = self.limit
         self.resolve_filters(
             model=self.model,
             q_objects=self.q_objects,
@@ -913,7 +992,7 @@ class DeleteQuery(AwaitableQuery):
 
     def __await__(self) -> Generator[Any, None, int]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db(True)  # type: ignore
         self._make_query()
         return self._execute().__await__()
 
@@ -951,7 +1030,7 @@ class ExistsQuery(AwaitableQuery):
 
     def __await__(self) -> Generator[Any, None, bool]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
         return self._execute().__await__()
 
@@ -993,7 +1072,7 @@ class CountQuery(AwaitableQuery):
 
     def __await__(self) -> Generator[Any, None, int]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
         return self._execute().__await__()
 
@@ -1187,7 +1266,7 @@ class ValuesListQuery(FieldSelectQuery):
 
     def __await__(self) -> Generator[Any, None, List[Any]]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
         return self._execute().__await__()  # pylint: disable=E1101
 
@@ -1273,7 +1352,7 @@ class ValuesQuery(FieldSelectQuery):
 
     def __await__(self) -> Generator[Any, None, List[dict]]:
         if self._db is None:
-            self._db = self.model._meta.db  # type: ignore
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
         return self._execute().__await__()  # pylint: disable=E1101